Improved the Makefile to be more modular
Refactored the Makefile so you can easily override the input directory for the SVG C4 diagram
files to generate them to Backstage catalog files to
the specified catalog output dir that you can easily specify now if
you want to.

Updated the README file with better clarification of the DrawWard-CLI usage.